pyproject.toml
[project]
name = "mcp-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.3",
    "google-adk[eval]>=1.23.0",
    "google-cloud-aiplatform[adk,agent-engines,evaluation]>=1.135.0",
    "langfuse>=3.12.1",
    "litellm>=1.81.5",
    "openinference-instrumentation-google-adk>=0.1.8",
]
